Move truncated distribution configs to delphi/utils/configs.py.
TruncatedExponentialFamilyDistributionConfig and
TruncatedMultivariateNormalConfig now live alongside TrainerConfig and
OptimizerConfig so any module can import them without pulling in the
full distribution implementation.
